Ensure CertIndex and lovelace types are consistent

+++ internal/datasource/cardano_db_sync/models/epoch_param.go
@@ -19,18 +19,18 @@ type EpochParam struct {
 	Entropy             []byte  `gorm:"column:entropy"` // This is a "hash32type" column
 	ProtocolMajor       uint32  `gorm:"column:protocol_major"`
 	ProtocolMinor       uint32  `gorm:"column:protocol_minor"`
-	MinUtxoValue        uint32  `gorm:"column:min_utxo_value"`      // This is a "lovelace" column
-	MinPoolCost         uint32  `gorm:"column:min_pool_cost"`       // This is a "lovelace" column
+	MinUtxoValue        uint64  `gorm:"column:min_utxo_value"`      // This is a "lovelace" column
+	MinPoolCost         uint64  `gorm:"column:min_pool_cost"`       // This is a "lovelace" column
 	Nonce               []byte  `gorm:"column:nonce"`               // This is a "hash32type" column
-	CoinsPerUtxoWord    uint32  `gorm:"column:coins_per_utxo_word"` // This is a "lovelace" column
+	CoinsPerUtxoWord    uint64  `gorm:"column:coins_per_utxo_word"` // This is a "lovelace" column
 	CostModelId         int64   `gorm:"column:cost_model_id"`       // cost_model(id)
 	PriceMem            float32 `gorm:"column:price_mem"`
 	PriceStep           float32 `gorm:"column:price_step"`
-	MaxTxExMem          string  `gorm:"column:max_tx_ex_mem"`      // This is a "word64type" column
-	MaxTxExSteps        string  `gorm:"column:max_tx_ex_steps"`    // This is a "word64type" column
-	MaxBlockExMem       string  `gorm:"column:max_block_ex_mem"`   // This is a "word64type" column
-	MaxBlockExSteps     string  `gorm:"column:max_block_ex_steps"` // This is a "word64type" column
-	MaxValSize          string  `gorm:"column:max_val_size"`       // This is a "word64type" column
+	MaxTxExMem          string  `gorm:"column:max_tx_ex_mem"`      // This is a "word63type" column
+	MaxTxExSteps        string  `gorm:"column:max_tx_ex_steps"`    // This is a "word63type" column
+	MaxBlockExMem       string  `gorm:"column:max_block_ex_mem"`   // This is a "word63type" column
+	MaxBlockExSteps     string  `gorm:"column:max_block_ex_steps"` // This is a "word63type" column
+	MaxValSize          string  `gorm:"column:max_val_size"`       // This is a "word63type" column
 	CollateralPercent   uint32  `gorm:"column:collateral_percent"`
 	MaxCollateralInputs uint32  `gorm:"column:max_collateral_inputs"`
 	BlockId             int64   `gorm:column:block_id"` // block(id)
